Makefile: move steamos-power-button rules/hwdb before 71-seat
After users updated to systemd v258, we received reports that the power
button was no longer working consistently in gamescope-session. After
debugging the issue with systemd maintainers, it seems we need to order
the rules file so it's placed before 71-seat.rules if we want to ensure
that we give the power button seat access along with `uaccess`.

Not sure how this worked on previous systemd versions, but this change
seems to fix the issue on systemd v258 while remaining backwards
compatible.

Tested with 75 reboot cycles on Steam Deck LCD, previously I could
reproduce the issue within 10 cycles.

Fixes: https://github.com/systemd/systemd/issues/39065